% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{getExperimentHubFilename}
\alias{getExperimentHubFilename}
\alias{tartare}
\title{getExperimentHubFilename}
\usage{
getExperimentHubFilename(filename)
}
\arguments{
\item{filename}{of the aws s3 blob.}
}
\value{
the file name of the local ExperimentHub.
}
\description{
getExperimentHubFilename
}
\examples{
fl <- system.file("extdata", "metadata.csv", package="tartare")   
metadata <- read.csv(fl, stringsAsFactors=FALSE)
     metadata$Title     
     
lapply(metadata$RDataPath, getExperimentHubFilename)
}
