% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{getFilename}
\alias{getFilename}
\alias{tartare}
\title{get packages files}
\usage{
getFilename(eh, query = c("tartare"))
}
\arguments{
\item{eh}{ExperimentHub object.}

\item{query}{a query string, e.g.,
\code{c('tartar', '20190710_003_PierceHeLaProteinDigestStd.raw')}}
}
\value{
returns file contained in the tartare package
}
\description{
calls the  \link[AnnotationHub]{query} method of an
\link[ExperimentHub]{ExperimentHub} 
objectquery and returns the local filenames.
Of note: for proprietary reasons, a symbolic link is set to the cached files.
}
\examples{
library(ExperimentHub)
eh <- ExperimentHub()
(files <- getFilename(eh))

\dontrun{
library(MsBackendRawFileReader)

be <- lapply(files, function(f){
if (grepl("mzXML$", f))
    backendInitialize(MsBackendMzR(), files = f)
else
    backendInitialize(MsBackendRawFileReader(), files = f, extra=FALSE)
})
}

\dontrun{

hfx.filename <- .query(eh, c('tartar', '20190710_003_PierceHeLaProteinDigestStd.raw'))
x <- .cnew ("Rawfile", hfx.filename)
x$GetInfoValues()
}
}
\author{
Christian Panse <cp@fgcz.ethz.ch>
}
